requirements.txt
gluonnlp==0.10.0
pandas==2.1.3
tqdm==4.66.1
sentencepiece==0.1.99
transformers==4.30.2
fastapi==0.112.2
uvicorn==0.20.0
attrdict==2.0.1
huggingface-hub==0.24.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
numpy==1.26.2
plotly==5.17.0
pydantic==1.10.13
regex==2023.10.3
requests==2.27.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
streamlit==1.28.2
tokenizers==0.13.3
torchmetrics==1.4.1
torchtext==0.15.2
torchvision==0.15.2
transformers==4.30.2
urllib3==1.26.18
virtualenv==15.1.0
wandb==0.15.12
transformers[torch]
langchain==0.2.8
langchain-chroma==0.1.2
langchain-community==0.2.7
langchain-core==0.2.19
langchain-openai==0.1.16
langchain-teddynote==0.0.28
langchain-text-splitters==0.2.2
langsmith==0.1.85
langchain-anthropic==0.1.20